feat(agent): make context_prune_threshold a fraction, default to 0.5 (#622)
* Make context_prune_threshold a fraction, default to 0.5

- Change ContextPruneThreshold from int (absolute tokens) to float64
  (fraction of usable input budget), defaulting to 0.5
- ContextPruneThresholdOrDefault() now clamps to min(prune, compaction)
  so pruning always runs before compaction regardless of config
- Add applyLegacyContextPruneThreshold() migration: detects values >= 1,
  logs a warning, and falls back to 0.5 so devices with the old absolute
  token count (e.g. 12000) still boot after upgrade
- Simplify historicalPruneBudgets(): single rule for both configured and
  automatic modes (trigger = budget × fraction, target = trigger × 6/7),
  collapsing four prune constants into one integer ratio
- Update docs, overlay comments, config-web metadata, and i18n (en/zh)

On a 10k budget: prune fires at 5000 (cleaning to 4285), compaction at
8000. With prune=0.9 and compaction=0.6, both land at 6000 and pruning
still runs first.

* Fix prune threshold tests that patched the default value

Both tests asserted an override using 0.5, which became the default in the
previous commit:

- TestUpdateConfigFileWritesIndependentContextPruneThreshold failed in CI.
  Service.Update filters a patch equal to the resolved current value as a
  no-op, so a 0.5 patch never reached the file and the write assertion
  failed.
- TestLoadRuntimeConfigCanOverrideContextPruneThreshold still passed but
  was vacuous for the same reason: it would have passed even if the
  configured override were ignored entirely.

Both now use 0.4 and note why the value must differ from the default.

* Reject non-finite context thresholds and fix stale doc wording

Addresses review findings on #622.

context_compaction_threshold = nan was accepted and stored raw: every
ordered comparison against NaN is false, so it passed both the < 0 and
>= 1 bounds. TOML accepts nan and inf literals into a float64 field, so
this was reachable from a hand-edited config, and the raw NaN then flowed
into the compaction trigger and into the config web JSON.

Behaviour for non-finite input was inconsistent three ways: prune = nan
and prune = inf were silently migrated to the default by the legacy
token-count migration, prune = -inf was rejected, and compaction = nan
was accepted. Now all six cases are rejected uniformly.

- Extract validateContextThresholdFraction so both fields share one rule,
  with an explicit NaN check.
- Stop the legacy migration from swallowing non-finite values; only a
  finite value >= 1 is a plausible token count, so NaN and Inf are left
  for Validate to reject rather than masked as the default.
- Fix the historicalPruneTarget comment, which still described the old
  0.7/0.6 default; state the current 0.5 -> 5000/4285 behaviour.
- Fix 'between 0 and 1' in the prune help text and agent.toml comment:
  0 is accepted (it selects the default) and 1 is not, so the inclusive
  phrasing was wrong in both directions. Updated in Go meta, both i18n
  locales, and their tests.

+++ src/agent/internal/configupdate/schema.go
@@ -475,7 +475,7 @@ type Agent struct {
 	Locale                     string  `json:"locale"`
 	CustomInstruction          string  `json:"custom_instruction"`
 	AdditionalPrompt           string  `json:"additional_prompt"`
-	ContextPruneThreshold      int     `json:"context_prune_threshold,omitempty"`
+	ContextPruneThreshold      float64 `json:"context_prune_threshold,omitempty"`
 	InputMode                  string  `json:"input_mode"`
 	VADBackend                 string  `json:"vad_backend"`
 	VADModelPath               string  `json:"vad_model_path"`
